refactor: extract DiagnosticsLogEntry.swift from BugNarratorDiagnostics.swift
Extracts DiagnosticsLogEntry struct (41 lines) from BugNarratorDiagnostics.swift
into its own file.

Byte-identical move; SHA-256 f00437bf58779fdc45a8daaff5d71cfb213fc344fb54e144ea5964fb08ae39e9
verified against post-#673 BugNarratorDiagnostics.swift lines 3-43.

Reduces BugNarratorDiagnostics.swift by 41 lines.

Tests: 667 (unchanged).
